Cast variables for windows.
git-svn-id: file:///home/jj/hercules.svn/trunk@6097 956126f8-22a0-4046-8f4a-272fa8102e63

+++ channel.c
@@ -3249,8 +3249,8 @@ resume_suspend:
     {
         dev->scsw.flag1 |= SCSW1_E;
         dev->esw.erw0 |= ERW0_S;
-        dev->esw.erw1 = (dev->numsense < (int)sizeof(dev->ecw)) ?
-                        dev->numsense : sizeof(dev->ecw);
+        dev->esw.erw1 = (BYTE)((dev->numsense < (int)sizeof(dev->ecw)) ?
+                        dev->numsense : (int)sizeof(dev->ecw));
         memcpy (dev->ecw, dev->sense, dev->esw.erw1 & ERW1_SCNT);
         memset (dev->sense, 0, sizeof(dev->sense));
         dev->sns_pending = 0;
